Refactor tree ordering into its own service
