+++ Meilidown/Interfaces/ITreeOrderService.cs
@@ -1,0 +1,9 @@
+﻿using Meilidown.Models.Index;
+
+namespace Meilidown.Interfaces
+{
+    public interface ITreeOrderService
+    {
+        Task ReorderTree(Dictionary<string, NavNode> tree);
+    }
+}
